requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
transformers==4.35.2
sentence-transformers==2.2.2
torch==2.1.1
scikit-learn==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1